aider: Adjust connection in `PtychoPINNModelParametersController` to use `valueChanged` signal of `DecimalLineEdit`.
# Aider chat conversation:

USER:
Traceback (most recent call last):
  File "/Users/ollie/anaconda3/envs/ptychodus/bin/ptychodus", line 5, in <module>
    from ptychodus.__main__ import main
  File "/Users/ollie/anaconda3/envs/ptychodus/lib/python3.10/site-packages/ptychodus/__main__.py", line 108, in <module>
    sys.exit(main())
  File "/Users/ollie/anaconda3/envs/ptychodus/lib/python3.10/site-packages/ptychodus/__main__.py", line 102, in main
    controller = ControllerCore.createInstance(model, view)
  File "/Users/ollie/anaconda3/envs/ptychodus/lib/python3.10/site-packages/ptychodus/controller/core.py", line 99, in createInstance
    controller = cls(model, view)
  File "/Users/ollie/anaconda3/envs/ptychodus/lib/python3.10/site-packages/ptychodus/controller/core.py", line 67, in __init__
    self._reconstructorParametersController = ReconstructorParametersController.createInstance(
  File "/Users/ollie/anaconda3/envs/ptychodus/lib/python3.10/site-packages/ptychodus/controller/reconstructor.py", line 72, in createInstance
    controller._addReconstructor(name)
  File "/Users/ollie/anaconda3/envs/ptychodus/lib/python3.10/site-packages/ptychodus/controller/reconstructor.py", line 108, in _addReconstructor
    widget = viewControllerFactory.createViewController(reconstructorName)
  File "/Users/ollie/anaconda3/envs/ptychodus/lib/python3.10/site-packages/ptychodus/controller/ptychopinn/factory.py", line 27, in createViewController
    controller = PtychoPINNParametersController.createInstance(self._model.modelPresenter,
  File "/Users/ollie/anaconda3/envs/ptychodus/lib/python3.10/site-packages/ptychodus/controller/ptychopinn/controller.py", line 25, in createInstance
    return cls(modelPresenter, trainingPresenter, view, fileDialogFactory)
  File "/Users/ollie/anaconda3/envs/ptychodus/lib/python3.10/site-packages/ptychodus/controller/ptychopinn/controller.py", line 16, in __init__
    self._modelParametersController = PtychoPINNModelParametersController.createInstance(
  File "/Users/ollie/anaconda3/envs/ptychodus/lib/python3.10/site-packages/ptychodus/controller/ptychopinn/model.py", line 29, in createInstance
    view.nPhotonsLineEdit.editingFinished.connect(lambda: presenter.setNPhotons(Decimal(view.nPhotonsLineEdit.text())))
AttributeError: 'DecimalLineEdit' object has no attribute 'editingFinished'

ASSISTANT: The error indicates that the `DecimalLineEdit` object does not have an `editingFinished` signal, which is typically available in `QLineEdit` objects. This suggests that the custom `DecimalLineEdit` widget does not directly expose the `editingFinished` signal or does not inherit from `QLineEdit` in a way that makes this signal available. To resolve this issue, you should connect the signal from the underlying `QLineEdit` if `DecimalLineEdit` is a composite widget or ensure that `DecimalLineEdit` correctly inherits from `QLineEdit` to expose the signal. Since I cannot modify the definition of `DecimalLineEdit` directly, I'll adjust the connection in the `PtychoPINNModelParametersController` to use a compatible signal or method.

Given the information provided, a suitable workaround is to use the `valueChanged` signal of `DecimalLineEdit` if it exists and is designed to emit signals similar to `QSpinBox` or another custom signal that serves a similar purpose. If `DecimalLineEdit` does not provide a `valueChanged` signal, you might need to implement it within the `DecimalLineEdit` class. Assuming `DecimalLineEdit` has a `valueChanged` signal or a method that can be used to trigger the update, here's how you can adjust the connection:

```diff
--- ptychodus/controller/ptychopinn/model.py
+++ ptychodus/controller/ptychopinn/model.py
@@ ... @@
-view.nPhotonsLineEdit.editingFinished.connect(lambda: presenter.setNPhotons(Decimal(view.nPhotonsLineEdit.text())))
+view.nPhotonsLineEdit.valueChanged.connect(lambda value: presenter.setNPhotons(value))
```

This change assumes that `valueChanged` emits the new value directly, which can be passed to `presenter.setNPhotons`. If `DecimalLineEdit` does not have a `valueChanged` signal but has another mechanism for detecting changes, you'll need to adjust the connection accordingly.